requirements.txt
Django==4.2.2
Pillow==9.5.0
python-decouple==3.8
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-import-export==3.2.0
django-jazzmin==2.6.0


